backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
redis[asyncio]==5.0.8
google-generativeai==0.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
supabase==2.9.1
# ML (조합 데이터셋 · XGBoost)
numpy>=1.26.0,<3
pandas>=2.1.0,<3
scikit-learn>=1.3.0,<2
xgboost>=2.0.0,<4
